backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"fastapi",
	"uvicorn"
]

[dependency-groups]
dev = [
	"ruff",
	"python-dotenv"
]
test = [
	"pytest",
	"pytest-asyncio",
	"httpx"
]

